feat(telegram): telegram bot integration
- added "Telegram" option as notification type (Reminders)
- backend classes refactoring
- feat(frontend): open modal detail with direct link
