+++ .env.example
@@ -41,6 +41,13 @@ NTFY_TOPIC_PREFIX=p-cal-
 NTFY_AUTH_TOKEN=your-ntfy-publisher-token
 # NTFY_AUTH_TOKEN is optional, needed for custom server with publishing auth
 
+# Telegram Bot Notifications
+TELEGRAM_ENABLED=false
+TELEGRAM_BOT_TOKEN=your-telegram-bot-token
+TELEGRAM_WEBHOOK_URL=
+# TELEGRAM_WEBHOOK_URL require a public HTTPS domain; If empty, polling method will be used instead (not recommended for production)
+TELEGRAM_USE_INLINE_BUTTONS=false # Enable this **ONLY IF the app is PUBLICLY ACCESSIBLE via HTTPS** or notifications will not works!
+# NOTE: Telegram does not permit notifications with action links who are not publicly accessible via HTTPS
 
 # ==============================================
 # FRONTEND CONFIGURATION (Vue.js)
