+++ docker-compose.yml
@@ -39,6 +39,10 @@ services:
       - NTFY_SERVER_URL=${NTFY_SERVER_URL}
       - NTFY_TOPIC_PREFIX=${NTFY_TOPIC_PREFIX}
       - NTFY_AUTH_TOKEN=${NTFY_AUTH_TOKEN}
+      - TELEGRAM_ENABLED=${TELEGRAM_ENABLED}
+      - TELEGRAM_BOT_TOKEN=${TELEGRAM_BOT_TOKEN}
+      - TELEGRAM_WEBHOOK_URL=${TELEGRAM_WEBHOOK_URL}
+      - TELEGRAM_USE_INLINE_BUTTONS=${TELEGRAM_USE_INLINE_BUTTONS:-false}
       - EMAIL_ENABLED=${EMAIL_ENABLED}
       - EMAIL_VERIFICATION=${EMAIL_VERIFICATION:-false}
       - MAIL_HOST=${MAIL_HOST}
