+++ frontend/src/components/TaskModal.vue
@@ -227,8 +227,9 @@
                     :disabled="isFormLoading"
                     class="w-full px-3 py-2 text-sm border border-gray-300 dark:border-gray-600 rounded bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:outline-none focus:ring-1 focus:ring-blue-500 disabled:opacity-50"
                   >
-                    <option value="PUSH">{{ t('tasks.reminderTypeNotification') }}</option>
+                    <option value="PUSH">{{ t('tasks.reminderTypeNtfy') }}</option>
                     <option value="EMAIL">{{ t('tasks.reminderTypeEmail') }}</option>
+                    <option value="TELEGRAM">{{ t('tasks.reminderTypeTelegram') }}</option>
                   </select>
                 </div>
