+++ frontend/src/i18n/locales/en-US.json
@@ -358,8 +358,9 @@
     "reminderType": "Type",
     "reminderValue": "Value",
     "reminderUnit": "Unit",
-    "reminderTypeNotification": "Notification",
+    "reminderTypeNtfy": "NTFY",
     "reminderTypeEmail": "Email",
+    "reminderTypeTelegram": "Telegram",
     "reminderUnitMinutes": "Minutes",
     "reminderUnitHours": "Hours",
     "reminderUnitDays": "Days",
