+++ frontend/src/i18n/locales/es-ES.json
@@ -365,8 +365,9 @@
     "reminderType": "Tipo",
     "reminderValue": "Valor",
     "reminderUnit": "Unidad",
-    "reminderTypeNotification": "Notificación",
+    "reminderTypeNtfy": "NTFY",
     "reminderTypeEmail": "Correo electrónico",
+    "reminderTypeTelegram": "Telegram",
     "reminderUnitMinutes": "Minutos",
     "reminderUnitHours": "Horas",
     "reminderUnitDays": "Días",
