+++ frontend/src/i18n/locales/it-IT.json
@@ -358,8 +358,9 @@
     "reminderType": "Tipologia",
     "reminderValue": "Valore",
     "reminderUnit": "Unità",
-    "reminderTypeNotification": "Notifica",
+    "reminderTypeNtfy": "NTFY",
     "reminderTypeEmail": "Email",
+    "reminderTypeTelegram": "Telegram",
     "reminderUnitMinutes": "Minuti",
     "reminderUnitHours": "Ore",
     "reminderUnitDays": "Giorni",
